requirements.txt
fastapi~=0.115.6
starlette~=0.41.3
pydantic~=2.10.3
python-dotenv~=1.0.1
sqlalchemy~=2.0.36
asyncpg~=0.30.0
